delete modified

+++ admin_area/delete_carousel.php
@@ -17,22 +17,30 @@ else {
 if(isset($_GET['delete_carousel'])){
 $delete_id = $_GET['delete_carousel'];
 
-$delete_com = "delete from carousel where id=".$delete_id;
+$sql = 'select image from carousel where id = '.$delete_id;
+$run = mysqli_query($con,$sql);
+$row = mysqli_fetch_array($run);
+$main_image = $row['image'];
+if (unlink("../storage/".$main_image))
+{
+    $delete_com = "delete from carousel where id=".$delete_id;
 
-$run_delete = mysqli_query($con,$delete_com);
+    $run_delete = mysqli_query($con,$delete_com);
 
-if($run_delete){
-    
-    
-    echo "<script>alert('One Image Has Been Removed')</script>";
-    
-    echo "<script>window.open('index.php?view_carousel','_self')</script>";
-    
+    if($run_delete)
+    {
+        echo "<script>alert('One Image Has Been Removed')</script>";
+        echo "<script>window.open('index.php?view_carousel','_self')</script>";
     }
     else
     {
         echo "<script>alert($con->error)</script>";
     }
+}
+else
+{
+    echo "<script>alert('Error Deleting File')</script>";
+}
     
 }
